+++ vm/ast/comp.h
@@ -2,7 +2,6 @@
 #if !defined(VM_HEADER_AST_COMP)
 #define VM_HEADER_AST_COMP
 
-#include "../ir/ir.h"
 #include "ast.h"
 
 void vm_ast_comp_more(vm_ast_node_t node, vm_blocks_t *blocks);
