+++ vm/backend/binop.inc
