+++ vm/io.h
@@ -8,7 +8,7 @@ typedef struct vm_io_debug_t vm_io_debug_t;
 struct vm_io_buffer_t;
 typedef struct vm_io_buffer_t vm_io_buffer_t;
 
-#include "../obj.h"
+#include "./obj.h"
 
 struct vm_io_debug_t {
     vm_io_debug_t *next;
