+++ vm/vm.h
@@ -8,15 +8,21 @@
 
 #define VM_FORMAT_FLOAT "%.14g"
 
+struct vm_t;
 struct vm_blocks_t;
-struct vm_config_t;
 struct vm_externs_t;
-struct vm_std_closure_t;
+struct vm_std_value_t;
+union vm_value_t;
+struct vm_table_t;
+struct vm_table_pair_t;
 
+typedef struct vm_t vm_t;
 typedef struct vm_blocks_t vm_blocks_t;
-typedef struct vm_config_t vm_config_t;
 typedef struct vm_externs_t vm_externs_t;
-typedef struct vm_std_closure_t vm_std_closure_t;
+typedef struct vm_std_value_t vm_std_value_t;
+typedef union vm_value_t vm_value_t;
+typedef struct vm_table_t vm_table_t;
+typedef struct vm_table_pair_t vm_table_pair_t;
 
 enum {
     VM_TAG_UNK,
@@ -48,18 +54,58 @@ enum {
 
 typedef uint8_t vm_tag_t;
 
+union vm_value_t {
+    void *all;
+    bool b;
+    int8_t i8;
+    int16_t i16;
+    int32_t i32;
+    int64_t i64;
+    float f32;
+    double f64;
+    const char *str;
+    vm_table_t *table;
+    vm_std_value_t *closure;
+    void (*ffi)(vm_t *closure, vm_std_value_t *args);
+};
+
+struct vm_std_value_t {
+    vm_value_t value;
+    vm_tag_t tag;
+};
+
+struct vm_table_pair_t {
+    vm_value_t key_val;
+    vm_value_t val_val;
+    vm_tag_t key_tag;
+    vm_tag_t val_tag;
+};
+
+struct vm_table_t {
+    vm_table_pair_t *pairs;
+    uint32_t len;
+    uint32_t used;
+    uint8_t alloc;
+};
+
 struct vm_externs_t {
     size_t id;
     void *value;
     vm_externs_t *last;
 };
 
-struct vm_config_t {
+struct vm_t {
     const char *save_file;
     vm_externs_t *externs;
     vm_blocks_t *blocks;
 
     uint8_t use_num;
+
+    vm_std_value_t std;
+
+    void *mutex;
+
+    vm_std_value_t *regs;
 };
 
 #endif
